+++ CHANGELOG.md
@@ -6,13 +6,15 @@ All notable changes to homebridge-dummy will be documented in this file.
 
 ### Added
 - `Doorbell` accessory type
-- History (Only visible with the [Eve app](https://apps.apple.com/us/app/eve-for-matter-homekit/id917695792))
+- History (Only visible with the [Eve app](https://apps.apple.com/us/app/eve-for-matter-homekit/id917695792)) ([documentation](https://github.com/mpatfield/homebridge-easy-mqtt#history))
     - `ContactSensorState`
     - `CurrentRelativeHumidity` (for `HumiditySensor`, `Thermostat`)
     - `CurrentTemperature` (for `HeaterCooler`, `TemperatureSensor`, `Thermostat`)
     - `MotionDetected`
     - `On` (for `Lightbulb`, `Outlet`, `Switch`)
-
+- Other topics and characteristics specific to the [Eve app](https://apps.apple.com/us/app/eve-for-matter-homekit/id917695792)
+    - `ContactSensor` - `TimesOpened`
+    - `Outlet` - `CurrentConsumption`, `TotalConsumption`, `ElectricCurrent`, `Voltage`
 - Traduzioni in italiano. Grazie, [@Shikaban](https://github.com/sponsors/Shikaban)!
 
 ### Notes